create generic SidebarListItem and wrap in NextLink
